Treat function types as unknown types for now
Previously they were treated as void* which resulted in function pointers
being interpreted as void** which doesn't make much sense.
It fixes exception when one address is defined both as function and as a
variable of function type at the same time (it is weird but can be found
in actual dwarf data from PS2).